// src/types/resources.ts
export type ForgetPasswordPageResourcesProps = {
  forgetPass: string;
  requiredEmailWithMinLengthOfFive: string;
  maxLength: string;
  character: string;
  emailNotValid: string;
  requiredPhone: string;
  phoneValidate: string;
  email: string;
  phone: string;
  resetPass: string;
  forgotPassPageDesc: string;
};